+++ next-docs/pages/components/label.tsx
@@ -1,24 +1,38 @@
 import React from 'react';
+import { Heading, Text } from '@heathmont/moon-core';
 import Preview from '../../components/codePreview/Preview';
 import PropsTable from '../../components/PropsTable';
 import Colours from '../../public/examples/label/Colours';
+import Icons from '../../public/examples/label/Icons';
+import LetterCases from '../../public/examples/label/LetterCases';
 import Sizes from '../../public/examples/label/Sizes';
-import SizesList from '../../public/examples/label/SizesList';
 import useExamples from '../../utils/useExamples';
 
 const PageLabel = () => {
   const examples = useExamples('label');
   return (
     <>
-      <section className="mt-8">
-        <h1 className="text-5xl font-semibold">Label</h1>
-        <p className="text-lg mt-4">Small count and labeling component.</p>
+      <section className="grid gap-4 mt-8">
+        <Heading as="h1" size={48}>
+          Label
+        </Heading>
+        <Text>
+          Use tags to label, categorize, or organize items using keywords that
+          describe them.
+        </Text>
+        <Text>Multiple or single tags can be used to categorize items.</Text>
+        <Text>
+          Use short labels for easy scanning. Use two words only if necessary to
+          describe the status and differentiate it from other tags.
+        </Text>
+        <Text>Default text style: Uppercase</Text>
       </section>
 
       {/* Colours */}
       <section className="mt-8">
         <Preview
           title="Colours"
+          isGrayBg
           preview={<Colours />}
           code={examples ? examples.Colours : 'Loading'}
         />
@@ -28,17 +42,29 @@ const PageLabel = () => {
       <section className="mt-8">
         <Preview
           title="Sizes"
+          isGrayBg
           preview={<Sizes />}
           code={examples ? examples.Sizes : 'Loading'}
         />
       </section>
 
-      {/* Sizes */}
+      {/* Icons */}
       <section className="mt-8">
         <Preview
-          title="Label with additional elements"
-          preview={<SizesList />}
-          code={examples ? examples.SizesList : 'Loading'}
+          title="Label with icons"
+          isGrayBg
+          preview={<Icons />}
+          code={examples ? examples.Icons : 'Loading'}
+        />
+      </section>
+
+      {/* Letter cases */}
+      <section className="mt-8">
+        <Preview
+          title="Letter cases"
+          isGrayBg
+          preview={<LetterCases />}
+          code={examples ? examples.LetterCases : 'Loading'}
         />
       </section>
 
@@ -47,10 +73,10 @@ const PageLabel = () => {
         data={[
           {
             name: 'size',
-            type: 'small | medium',
+            type: 'twoxsmall | xsmall',
             required: false,
-            default: 'medium',
-            description: 'Small/medium size for label',
+            default: 'xsmall',
+            description: 'Size for label',
           },
           {
             name: 'iconLeft',
@@ -64,13 +90,13 @@ const PageLabel = () => {
             type: 'React.ReactElement',
             required: false,
             default: '-',
-            description: 'Rigth text element',
+            description: 'Right text element',
           },
           {
             name: 'color',
             type: 'ColorProps',
             required: false,
-            default: 'goten.100',
+            default: 'gohan.100',
             description: 'Text color',
           },
           {
@@ -79,6 +105,13 @@ const PageLabel = () => {
             required: false,
             default: 'bulma.100',
             description: 'Background color',
+          },
+          {
+            name: 'isUppercase',
+            type: 'boolean',
+            required: false,
+            default: 'true',
+            description: 'Letter case',
           },
         ]}
       />