+++ next-docs/public/examples/avatar/Colours.tsx
@@ -2,8 +2,9 @@ import { Avatar } from '@heathmont/moon-components';
 
 const Example = () => (
   <div className="flex justify-around items-center w-full">
-    <Avatar color="gohan.100" backgroundColor="piccolo.100" />
-    <Avatar color="piccolo.100" backgroundColor="gohan.100" />
+    <Avatar />
+    <Avatar color="goten.100" backgroundColor="piccolo.100" />
+    <Avatar color="goten.100" backgroundColor="chiChi.100" />
   </div>
 );