+++ next-docs/public/examples/label/Colours.tsx
@@ -3,9 +3,9 @@ import { Label } from '@heathmont/moon-core';
 
 const Example = () => (
   <div className="flex justify-around items-center w-full">
-    <Label>Medium</Label>
-    <Label color="piccolo.100" backgroundColor="trunks.100">
-      Medium
+    <Label>Xsmall</Label>
+    <Label color="goten.100" backgroundColor="trunks.100">
+      Xsmall
     </Label>
     <Label color="goten.100" backgroundColor="dodoria.100">
       Medium