+++ packages/core/src/label/styles/LabelContainer.ts
@@ -1,30 +1,29 @@
-import { ColorProps } from '@heathmont/moon-themes';
 import { rem, themed } from '@heathmont/moon-utils';
 import styled from 'styled-components';
+import setFontSize from '../utils/setFontSize';
+import setLetterSpacing from '../utils/setLetterSpacing';
+import setPadding from '../utils/setPadding';
+import type LabelProps from '../private/types/LabelProps';
 
-type Props = {
-  color?: ColorProps;
-  backgroundColor?: ColorProps;
-  size?: 'small' | 'medium';
-};
-
-const LabelContainer = styled.span<Props>(
+const LabelContainer = styled.span<LabelProps>(
   ({ theme: { newTokens } }) => ({
     display: 'flex',
     alignItems: 'center',
-    fontWeight: newTokens.font.weight.semibold,
-    textTransform: 'uppercase',
-    borderRadius: newTokens.borderRadius.full,
-    letterSpacing: rem(1),
-    fontSize: rem(10),
+    borderRadius: newTokens.borderRadius.interactive,
     lineHeight: rem(16),
   }),
-  ({ size }) => ({
-    padding: size === 'small' ? `0 ${rem(8)}` : `${rem(2)} ${rem(8)}`,
+  ({ size, iconLeft, iconRight }) => ({
+    padding: setPadding({ size, iconRight, iconLeft }),
+  }),
+  ({ isUppercase, size, theme: { newTokens } }) => ({
+    textTransform: isUppercase ? 'uppercase' : 'none',
+    fontWeight: isUppercase && newTokens.font.weight.semibold,
+    letterSpacing: setLetterSpacing({ isUppercase, size }),
+    fontSize: setFontSize({ isUppercase, size }),
   }),
   /* If a color or backgroundColor is set, override the modifier styles. */
   ({ color, backgroundColor, theme }) => ({
-    color: color ? themed('color', color)(theme) : theme.colorNew.goten,
+    color: color ? themed('color', color)(theme) : theme.colorNew.gohan,
     backgroundColor: backgroundColor
       ? themed('color', backgroundColor)(theme)
       : theme.colorNew.bulma,